requirements.txt
beautifulsoup4==4.9.3
boto3==1.17.33
botocore==1.20.33
bs4==0.0.1
certifi==2020.12.5
chardet==4.0.0
idna==2.10
jmespath==0.10.0
numpy==1.20.1
Pillow @ file:///Users/nandom/Downloads/Pillow-8.0.1-cp38-cp38-macosx_11_0_arm64.whl
python-dateutil==2.8.1
pytz==2021.1
requests==2.25.1
s3transfer==0.3.6
six==1.15.0
soupsieve==2.2.1
urllib3==1.26.4
